+++ misskey-api/src/endpoint/channels.rs
@@ -7,3 +7,15 @@ pub mod show;
 pub mod timeline;
 pub mod unfollow;
 pub mod update;
+
+#[cfg(feature = "13-11-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
+pub mod favorite;
+
+#[cfg(feature = "13-11-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
+pub mod my_favorites;
+
+#[cfg(feature = "13-11-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
+pub mod unfavorite;