+++ misskey-api/src/endpoint/channels/update.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "13-11-0")]
+use crate::model::note::Note;
 use crate::model::{channel::Channel, drive::DriveFile, id::Id};
 
 use serde::Serialize;
@@ -19,6 +21,11 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub banner_id: Option<Option<Id<DriveFile>>>,
+    #[cfg(feature = "13-11-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-11-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub pinned_note_ids: Option<Vec<Id<Note>>>,
 }
 
 impl misskey_core::Request for Request {
@@ -48,6 +55,8 @@ mod tests {
             name: Some("yCdKKHkRAmqhE49j3TBCVnnsQiZ2KkCK83z6NNKoGaiqQNOALsAOIz6XVJAcaV9lNbQYuuhSe7nAM8qdvUtokhWYWDO999z07N4ajtikDmzANpgwRh7rxMgb8PLsgcbm".to_string()),
             description: None,
             banner_id: None,
+            #[cfg(feature = "13-11-0")]
+            pinned_note_ids: None,
         }).await;
     }
 
@@ -68,6 +77,8 @@ mod tests {
                 name: None,
                 description: Some(Some("description".to_string())),
                 banner_id: None,
+                #[cfg(feature = "13-11-0")]
+                pinned_note_ids: None,
             })
             .await;
         client
@@ -76,6 +87,8 @@ mod tests {
                 name: None,
                 description: Some(None),
                 banner_id: None,
+                #[cfg(feature = "13-11-0")]
+                pinned_note_ids: None,
             })
             .await;
     }
@@ -99,6 +112,8 @@ mod tests {
                 name: None,
                 description: None,
                 banner_id: Some(Some(file.id)),
+                #[cfg(feature = "13-11-0")]
+                pinned_note_ids: None,
             })
             .await;
         // bug in misskey
@@ -110,5 +125,29 @@ mod tests {
         //         banner_id: Some(None),
         //     })
         //     .await;
+    }
+
+    #[cfg(feature = "13-11-0")]
+    #[tokio::test]
+    async fn request_with_pinned_note_ids() {
+        let client = TestClient::new();
+        let channel = client
+            .test(crate::endpoint::channels::create::Request {
+                name: "test channel".to_string(),
+                description: None,
+                banner_id: None,
+            })
+            .await;
+        let note = client.create_note(Some("test"), None, None).await;
+
+        client
+            .test(Request {
+                channel_id: channel.id,
+                name: None,
+                description: None,
+                banner_id: None,
+                pinned_note_ids: Some(vec![note.id]),
+            })
+            .await;
     }
 }