+++ misskey-api/src/endpoint/i/notifications.rs
@@ -22,11 +22,13 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub until_id: Option<Id<Notification>>,
+    #[cfg(not(feature = "13-11-0"))]
+    #[cfg_attr(docsrs, doc(cfg(not(feature = "13-11-0"))))]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub following: Option<bool>,
-    #[cfg(feature = "12-92-0")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "12-92-0")))]
+    #[cfg(all(feature = "12-92-0", not(feature = "13-11-0")))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "12-92-0", not(feature = "13-11-0")))))]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub unread_only: Option<bool>,
@@ -67,8 +69,9 @@ mod tests {
                 limit: Some(100),
                 since_id: None,
                 until_id: None,
+                #[cfg(not(feature = "13-11-0"))]
                 following: None,
-                #[cfg(feature = "12-92-0")]
+                #[cfg(all(feature = "12-92-0", not(feature = "13-11-0")))]
                 unread_only: None,
                 mark_as_read: None,
                 include_types: None,
@@ -87,8 +90,9 @@ mod tests {
                 limit: None,
                 since_id: None,
                 until_id: None,
+                #[cfg(not(feature = "13-11-0"))]
                 following: Some(true),
-                #[cfg(feature = "12-92-0")]
+                #[cfg(all(feature = "12-92-0", not(feature = "13-11-0")))]
                 unread_only: Some(true),
                 mark_as_read: Some(false),
                 include_types: Some(
@@ -120,8 +124,9 @@ mod tests {
                     limit: None,
                     since_id: None,
                     until_id: None,
+                    #[cfg(not(feature = "13-11-0"))]
                     following: None,
-                    #[cfg(feature = "12-92-0")]
+                    #[cfg(all(feature = "12-92-0", not(feature = "13-11-0")))]
                     unread_only: None,
                     mark_as_read: None,
                     include_types: None,
@@ -137,8 +142,9 @@ mod tests {
                 limit: None,
                 since_id: Some(notification_id.clone()),
                 until_id: Some(notification_id.clone()),
+                #[cfg(not(feature = "13-11-0"))]
                 following: None,
-                #[cfg(feature = "12-92-0")]
+                #[cfg(all(feature = "12-92-0", not(feature = "13-11-0")))]
                 unread_only: None,
                 mark_as_read: None,
                 include_types: None,