+++ misskey-api/src/endpoint/notifications.rs
@@ -4,6 +4,6 @@ pub mod create;
 
 pub mod mark_all_as_read;
 
-#[cfg(feature = "12-77-1")]
-#[cfg_attr(docsrs, doc(cfg(feature = "12-77-1")))]
+#[cfg(all(feature = "12-77-1", not(feature = "13-11-0")))]
+#[cfg_attr(docsrs, doc(cfg(all(feature = "12-77-1", not(feature = "13-11-0")))))]
 pub mod read;